trawler/requirements.txt
bs4
pyyaml
requests
wget
tldextract
pika